Development for v1.4.7

+++ source/debug_funcs.hpp
@@ -11,8 +11,8 @@
  *   Note: Please be aware that this notice cannot be altered or removed. It is a part
  *   of the project's documentation and must remain intact.
  * 
+ *  Licensed under CC BY-NC-SA 4.0
  *  Copyright (c) 2023 ppkantorski
- *  All rights reserved.
  ********************************************************************************/
 
 #pragma once
@@ -48,7 +48,7 @@ void trimLog(FILE* file) {
             // Split the buffer into lines
             char* line = strtok(buffer, "\n");
             bool firstLine = true;
-            size_t lineCount = 0;
+            int lineCount = 0;
             
             while (line != nullptr) {
                 if (!firstLine) {
